Refs #255 (2.6) Added description for select Def scene. But cannot get fill height to work correctly.

+++ game/src/Voider/src/com/spiddekauga/voider/game/Level.java
@@ -102,7 +102,7 @@ public class Level extends Resource implements Disposable, IResourceRevision {
 	}
 
 	@Override
-	public long getRevision() {
+	public int getRevision() {
 		return mLevelDef.getRevision();
 	}
